+++ CHANGELOG.md
@@ -38,6 +38,8 @@
 
 ### Added
 
+- nostr: add NIP-88 support ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/892)
+
 ### Fixed
 
 ### Removed